Core 4.3.2

+++ docker-entrypoint.sh
@@ -1,12 +1,11 @@
 #!/bin/bash
-cp -r /usr/local/nagios/etc-ori/* /usr/local/nagios/etc/
+if [ ! -f "/usr/local/nagios/etc/nagios.cfg" ] 
+then
+    echo "Copying files for first start to volume" 
+    cp -r /usr/local/nagios/etc-ori/* /usr/local/nagios/etc/
+    echo "Create defualt httpasswd. Log into terminal and run: 'htpasswd -bc /usr/local/nagios/etc/htpasswd.users ${NAGIOS_WEBADMIN_USER} <yourpassword>'" 
+    htpasswd -bc /usr/local/nagios/etc/htpasswd.users ${NAGIOS_WEBADMIN_USER} ${NAGIOS_WEBADMIN_START_PASSWORD}
+fi
 
-# set -e
-# if [ "$1" = 'postgres' ]; then
-#    chown -R postgres "$PGDATA"
-#    if [ -z "$(ls -A "$PGDATA")" ]; then
-#        gosu postgres initdb
-#    fi
-#    exec gosu postgres "$@"
-# fi
-# exec "$@"
+echo "Starting supervisord ..." 
+exec /usr/bin/supervisord --nodaemon --configuration /tmp/supervisor_nagios.conf
